refactor: restore leaf seams in daemon cleanup
