+++ src/daemon-client.ts
@@ -3,7 +3,6 @@ import http from 'node:http';
 import https from 'node:https';
 import fs from 'node:fs';
 import path from 'node:path';
-import { spawn } from 'node:child_process';
 import { AppError } from './utils/errors.ts';
 import type {
   DaemonArtifact,
@@ -22,6 +21,7 @@ import {
   type DaemonServerMode,
   type DaemonTransportPreference,
 } from './daemon/config.ts';
+import { uploadArtifact } from './upload-client.ts';
 export type DaemonRequest = SharedDaemonRequest;
 export type DaemonResponse = SharedDaemonResponse;
 
@@ -865,7 +865,9 @@ async function sendSocketRequest(info: DaemonInfo, req: DaemonRequest): Promise<
     const socket = net.createConnection({ host: '127.0.0.1', port }, () => {
       socket.write(`${JSON.stringify(req)}\n`);
     });
-    const statePaths = resolveDaemonPaths(req.flags?.stateDir ?? process.env.AGENT_DEVICE_STATE_DIR);
+    const statePaths = resolveDaemonPaths(
+      req.flags?.stateDir ?? process.env.AGENT_DEVICE_STATE_DIR,
+    );
     const timeout = setTimeout(() => {
       socket.destroy();
       reject(handleRequestTimeout(info, statePaths, req.meta?.requestId, req.command, false));
@@ -1287,124 +1289,4 @@ export function resolveDaemonStartupHint(
     return `Daemon metadata may be stale. If no agent-device daemon process is running, delete ${paths.infoPath} and ${paths.lockPath}, then retry.`;
   }
   return `Daemon metadata is missing or stale. Delete ${paths.infoPath} if present and retry.`;
-}
-
-const UPLOAD_TIMEOUT_MS = 5 * 60 * 1000; // 5 minutes
-
-type UploadArtifactOptions = {
-  localPath: string;
-  baseUrl: string;
-  token: string;
-};
-
-type UploadResponse = {
-  ok: boolean;
-  uploadId: string;
-};
-
-function uploadArtifact(options: UploadArtifactOptions): Promise<string> {
-  const { localPath, baseUrl, token } = options;
-
-  const stat = fs.statSync(localPath);
-  const isDirectory = stat.isDirectory();
-  const filename = path.basename(localPath);
-  const artifactType = isDirectory ? 'app-bundle' : 'file';
-
-  const normalizedBase = baseUrl.endsWith('/') ? baseUrl : `${baseUrl}/`;
-  const uploadUrl = new URL('upload', normalizedBase);
-  const transport = uploadUrl.protocol === 'https:' ? https : http;
-
-  const headers: Record<string, string> = {
-    'x-artifact-type': artifactType,
-    'x-artifact-filename': filename,
-    'transfer-encoding': 'chunked',
-  };
-  if (token) {
-    headers.authorization = `Bearer ${token}`;
-    headers['x-agent-device-token'] = token;
-  }
-
-  return new Promise((resolve, reject) => {
-    const req = transport.request(
-      {
-        protocol: uploadUrl.protocol,
-        host: uploadUrl.hostname,
-        port: uploadUrl.port,
-        method: 'POST',
-        path: uploadUrl.pathname + uploadUrl.search,
-        headers,
-      },
-      (res) => {
-        let body = '';
-        res.setEncoding('utf8');
-        res.on('data', (chunk) => {
-          body += chunk;
-        });
-        res.on('end', () => {
-          clearTimeout(timeout);
-          try {
-            const parsed = JSON.parse(body) as UploadResponse;
-            if (!parsed.ok || !parsed.uploadId) {
-              reject(new AppError('COMMAND_FAILED', `Upload failed: ${body}`));
-              return;
-            }
-            resolve(parsed.uploadId);
-          } catch {
-            reject(new AppError('COMMAND_FAILED', `Invalid upload response: ${body}`));
-          }
-        });
-      },
-    );
-
-    const timeout = setTimeout(() => {
-      req.destroy();
-      reject(
-        new AppError('COMMAND_FAILED', 'Artifact upload timed out', {
-          timeoutMs: UPLOAD_TIMEOUT_MS,
-          hint: 'The upload to the remote daemon exceeded the 5-minute timeout.',
-        }),
-      );
-    }, UPLOAD_TIMEOUT_MS);
-
-    req.on('error', (err) => {
-      clearTimeout(timeout);
-      reject(
-        new AppError(
-          'COMMAND_FAILED',
-          'Failed to upload artifact to remote daemon',
-          { hint: 'Verify the remote daemon is reachable and supports artifact uploads.' },
-          err,
-        ),
-      );
-    });
-
-    if (isDirectory) {
-      const parentDir = path.dirname(localPath);
-      const dirName = path.basename(localPath);
-      const tar = spawn('tar', ['cf', '-', '-C', parentDir, dirName], {
-        stdio: ['ignore', 'pipe', 'pipe'],
-      });
-      tar.stdout.pipe(req);
-      tar.on('error', (err) => {
-        req.destroy();
-        reject(
-          new AppError('COMMAND_FAILED', 'Failed to create tar archive for app bundle', {}, err),
-        );
-      });
-      tar.on('close', (code) => {
-        if (code !== 0) {
-          req.destroy();
-          reject(new AppError('COMMAND_FAILED', `tar failed with exit code ${code}`));
-        }
-        // tar stdout end will trigger req.end() via pipe
-      });
-    } else {
-      const fileStream = fs.createReadStream(localPath);
-      fileStream.pipe(req);
-      fileStream.on('error', (err) => {
-        req.destroy();
-        reject(new AppError('COMMAND_FAILED', 'Failed to read local artifact', {}, err));
-      });
-    }
-  });
 }
