+++ src/utils/merge-flags.ts
@@ -1,4 +1,7 @@
-export function mergeDefinedFlags<T extends Record<string, unknown>>(target: T, source: Partial<T>): T {
+export function mergeDefinedFlags<T extends Record<string, unknown>>(
+  target: T,
+  source: Partial<T>,
+): T {
   for (const [key, value] of Object.entries(source)) {
     if (value !== undefined) {
       target[key as keyof T] = value as T[keyof T];
